(refactor) Depr common editWorkoutComponents and rename common viewWorkoutComponents to workoutViews
