+++ client/scripts/common/workoutViews/exerciseDescrText.js
