+++ client/scripts/common/workoutViews/part.js
