+++ client/scripts/common/workoutViews/renderResultsTime.js
