+++ client/scripts/common/workoutViews/viewWorkoutBody.js
