+++ client/scripts/components/feed/feedCard.js
@@ -2,7 +2,7 @@
 * @Author: vincetam
 * @Date:   2015-07-30 12:52:52
 * @Last Modified by:   VINCE
-* @Last Modified time: 2016-01-11 14:40:25
+* @Last Modified time: 2016-01-19 09:49:11
 */
 
 'use strict';
@@ -12,8 +12,8 @@ var feedActions = require('../../actions/feedActions');
 
 //Load components
 var FeedCardHeader = require('./feedCardHeader');
-var ViewWorkoutBody = require('../../common/viewWorkoutComponents/viewWorkoutBody');
-var ViewResults = require('../../common/viewWorkoutComponents/viewResults');
+var ViewWorkoutBody = require('../../common/workoutViews/viewWorkoutBody');
+var ViewResults = require('../../common/workoutViews/viewResults');
 var FeedCardFooter = require('./feedCardFooter');
 
 var {
